+++ src/vector3d_utils2.c
@@ -6,12 +6,19 @@
 /*   By: tterao <tterao@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 13:48:29 by ryhara            #+#    #+#             */
-/*   Updated: 2023/10/09 17:16:25 by tterao           ###   ########.fr       */
+/*   Updated: 2023/10/11 16:10:47 by tterao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector3d.h"
 
+/**
+ * @brief 外積を計算する
+ *
+ * @param v1
+ * @param v2
+ * @return t_vector3d
+ */
 t_vector3d	vector3d_cross(const t_vector3d v1, const t_vector3d v2)
 {
 	t_vector3d	result;
@@ -22,7 +29,7 @@ t_vector3d	vector3d_cross(const t_vector3d v1, const t_vector3d v2)
 	return (result);
 }
 
-t_vector3d	vector3d_create(float x, float y, float z)
+t_vector3d	vector3d_create(double x, double y, double z)
 {
 	t_vector3d	new;
